updated gateway to remove team members when no longer in team
